refactor: convert class method to standalone function
(cherry picked from commit 1ea1bfebc4a2407961d93a6d0c4c6c9f43202689)
